added and configured tsconfig, fixed up some broken routes and modified fireBase admin to work in the new setting. Restructured the backend folder to work with the directory
