+++ backend/src/middleware/authMiddleware.ts
