+++ backend/src/routes/adminRoutes/viewGrades.ts
@@ -1,8 +1,8 @@
 import "dotenv/config";
 import { Router } from "express";
-import { PrismaClient } from "../../generated/prisma-client/client.ts";
+import { PrismaClient } from "../../generated/prisma-client/client";
 import { z } from "zod";
-import { subjectSchema, yearSchema } from "../../validators/valdation.js";
+import { subjectSchema, yearSchema } from "../../validators/valdation";
 
 const prisma = new PrismaClient();
 const router = Router();
