+++ backend/src/routes/studentRoutes.ts
@@ -1,6 +1,6 @@
 import "dotenv/config";
 import { Router } from "express";
-import { PrismaClient } from "../generated/prisma-client/client.ts";
+import { PrismaClient } from "../generated/prisma-client/client";
 import { z } from "zod";
 import { studentSchema } from "../validators/valdation.js";
